requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.12
botocore==1.19.38
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-currentuser==0.5.2
django-forms-bootstrap==3.1.0
django-widget-tweaks==1.4.8
docker==4.4.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
numpy==1.19.4
paramiko==2.7.2
pathspec==0.5.9
payjp==0.0.5
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
urllib3==1.25.11
virtualenv==20.2.1
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==3.3.1
